fix(Inline): Remove top margin when Inline is not first child
