+++ lib/hooks/useNegativeMargin/useNegativeMarginLeft.ts
@@ -4,14 +4,16 @@ import {
   resolveResponsiveProp,
   ResponsiveProp,
 } from '../../utils/responsiveProp';
-import * as styleRefs from './useNegativeOffsetX.treat';
+import * as styleRefs from './useNegativeMarginLeft.treat';
 
-type NegativeOffset = Extract<
+type NegativeMarginLeft = Extract<
   Extract<keyof typeof styleRefs.mobile, keyof typeof styleRefs.tablet>,
   Extract<keyof typeof styleRefs.mobile, keyof typeof styleRefs.desktop>
 >;
 
-export const useNegativeOffsetX = (space: ResponsiveProp<NegativeOffset>) => {
+export const useNegativeMarginLeft = (
+  space: ResponsiveProp<NegativeMarginLeft>,
+) => {
   const styles = useStyles(styleRefs);
   return classnames(
     resolveResponsiveProp(space, styles.mobile, styles.tablet, styles.desktop),
