--- lib/hooks/useNegativeOffset/useNegativeOffset.ts
@@ -1,2 +1,0 @@
-export { useNegativeOffsetX } from './useNegativeOffsetX';
-export { useNegativeOffsetY } from './useNegativeOffsetY';
